homework_1/requirements.txt
anyio==3.6.1
asyncpg==0.26.0
click==8.1.3
fastapi==0.80.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
pydantic==1.9.2
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
python-jose==3.3.0
alembic==1.8.1
pydantic[dotenv]